+++ src/test/java/com/epam/reportportal/junit/ignore/IgnoreParameterizedTest.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package com.epam.reportportal.junit.retry;
+package com.epam.reportportal.junit.ignore;
 
 import com.epam.reportportal.junit.ParallelRunningHandler;
-import com.epam.reportportal.junit.features.retry.StandardParametersRetryTest;
+import com.epam.reportportal.junit.features.skip.ParameterizedIgnoredTest;
 import com.epam.reportportal.junit.utils.TestUtils;
 import com.epam.reportportal.listeners.ItemStatus;
 import com.epam.reportportal.service.ReportPortal;
@@ -38,12 +38,12 @@ import static org.hamcrest.Matchers.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
-public class ParameterizedRetryTest {
+public class IgnoreParameterizedTest {
 
 	private final String launchId = CommonUtils.namedId("launch_");
 	private final String suiteId = CommonUtils.namedId("suite_");
 	private final String classId = CommonUtils.namedId("class_");
-	private final List<String> methodIds = Stream.generate(() -> CommonUtils.namedId("method_")).limit(3).collect(Collectors.toList());
+	private final List<String> methodIds = Stream.generate(() -> CommonUtils.namedId("method_")).limit(2).collect(Collectors.toList());
 
 	private final ReportPortalClient client = mock(ReportPortalClient.class);
 
@@ -55,32 +55,27 @@ public class ParameterizedRetryTest {
 	}
 
 	@Test
-	public void verify_a_parameterized_test_with_one_retry_for_the_first_parameter() {
-		TestUtils.runClasses(StandardParametersRetryTest.class);
+	public void verify_a_test_with_one_retry() {
+		TestUtils.runClasses(ParameterizedIgnoredTest.class);
 
 		verify(client, times(1)).startTestItem(any());
 		ArgumentCaptor<StartTestItemRQ> startCaptor = ArgumentCaptor.forClass(StartTestItemRQ.class);
 		ArgumentCaptor<FinishTestItemRQ> finishCaptor = ArgumentCaptor.forClass(FinishTestItemRQ.class);
 		verify(client, times(1)).startTestItem(same(suiteId), any());
-		verify(client, times(3)).startTestItem(same(classId), startCaptor.capture());
-		verify(client, times(2)).finishTestItem(same(methodIds.get(0)), finishCaptor.capture());
+		verify(client, times(2)).startTestItem(same(classId), startCaptor.capture());
+		verify(client, times(1)).finishTestItem(same(methodIds.get(0)), finishCaptor.capture());
 		verify(client, times(1)).finishTestItem(same(methodIds.get(1)), finishCaptor.capture());
-		verify(client, times(1)).finishTestItem(same(methodIds.get(2)), finishCaptor.capture());
 
-		List<StartTestItemRQ> startItems = startCaptor.getAllValues();
-		assertThat(startItems.get(0).isRetry(), nullValue());
-		assertThat(startItems.get(1).isRetry(), allOf(notNullValue(), equalTo(Boolean.TRUE)));
-		assertThat(startItems.get(1).getTestCaseId(), allOf(notNullValue(), equalTo(startItems.get(0).getTestCaseId())));
-		assertThat(startItems.get(2).isRetry(), nullValue());
+		List<StartTestItemRQ> startRqs = startCaptor.getAllValues();
+		assertThat(startRqs.get(0).getTestCaseId(), not(equalTo(startRqs.get(1).getTestCaseId()))); // test case ID should be parameterized
+		assertThat(startRqs.get(0).getParameters(), allOf(notNullValue(), hasSize(1)));
+		assertThat(startRqs.get(1).getParameters(), allOf(notNullValue(), hasSize(1)));
+		assertThat(startRqs.get(0).getCodeRef(), equalTo(startRqs.get(1).getCodeRef()));
 
-		List<FinishTestItemRQ> finishItems = finishCaptor.getAllValues();
-		assertThat(finishItems.get(0).isRetry(), nullValue());
-		assertThat(finishItems.get(1).isRetry(), allOf(notNullValue(), equalTo(Boolean.TRUE)));
-		assertThat(finishItems.get(1).getStatus(), allOf(notNullValue(), equalTo(finishItems.get(0).getStatus())));
-		assertThat(finishItems.get(2).isRetry(), nullValue());
-		assertThat(finishItems.get(2).getStatus(), allOf(notNullValue(), equalTo(ItemStatus.PASSED.name())));
-		assertThat(finishItems.get(3).isRetry(), nullValue());
-		assertThat(finishItems.get(3).getStatus(), allOf(notNullValue(), equalTo(ItemStatus.PASSED.name())));
+		finishCaptor.getAllValues().forEach(i -> {
+			assertThat(i.getIssue(), nullValue());
+			assertThat(i.getStatus(), equalTo(ItemStatus.SKIPPED.name()));
+		});
 	}
 
 }